make file, C++ class and model names consistent for synapses
