+++ models/clopath_synapse.h
@@ -1,5 +1,5 @@
 /*
- *  urbanczik_connection.h
+ *  clopath_synapse.h
  *
  *  This file is part of NEST.
  *
@@ -20,8 +20,8 @@
  *
  */
 
-#ifndef URBANCZIK_CONNECTION_H
-#define URBANCZIK_CONNECTION_H
+#ifndef CLOPATH_SYNAPSE_H
+#define CLOPATH_SYNAPSE_H
 
 // C++ includes:
 #include <cmath>
@@ -40,53 +40,67 @@
 namespace nest
 {
 
-/** @BeginDocumentation
-@ingroup Synapses
-@ingroup stdp
+/* BeginUserDocs: synapse, spike-timing-dependent plasticity, Clopath plasticity
 
-Name: urbanczik_synapse - Synapse type for a plastic synapse after Urbanczik and Senn.
+Short description
++++++++++++++++++
 
-Description:
+Synapse type for voltage-based STDP after Clopath
 
-urbanczik_synapse is a connector to create Urbanczik synapses as defined in [1] that can connect suitable
-multicompartment models. In contrast to usual STDP, the change of the synaptic weight does not only depend on the pre-
-and postsynaptic spike timing but also on the postsynaptic dendritic potential.
+Description
++++++++++++
 
-Urbanczik synapses require archiving of continuous quantities. Therefore they can only be connected to neuron models
-that are capable of doing this archiving. So far, the only compatible model is pp_cond_exp_mc_urbanczik.
+clopath_synapse is a connector to create Clopath synapses as defined
+in [1]_. In contrast to usual STDP, the change of the synaptic weight does
+not only depend on the pre- and postsynaptic spike timing but also on the
+postsynaptic membrane potential.
 
-Parameters:
+Clopath synapses require archiving of continuous quantities. Therefore Clopath
+synapses can only be connected to neuron models that are capable of doing this
+archiving. So far, compatible models are aeif_psc_delta_clopath and
+hh_psc_alpha_clopath.
 
-\verbatim embed:rst
-=======   ======  ==========================================================
-eta       real    Learning rate
-tau_Delta real    Time constant of low pass filtering of the weight change
-Wmax      real    Maximum allowed weight
-Wmin      real    Minimum allowed weight
-=======   ======  ==========================================================
-\endverbatim
+Parameters
+++++++++++
 
-All other parameters are stored in in the neuron models that are compatible with the Urbanczik synapse.
+=======  ======  ==========================================================
+tau_x    ms      Time constant of the trace of the presynaptic spike train
+Wmax     real    Maximum allowed weight
+Wmin     real    Minimum allowed weight
+=======  ======  ==========================================================
 
-Remarks:
+Other parameters like the amplitudes for long-term potentiation (LTP) and
+depression (LTD) are stored in in the neuron models that are compatible with the
+Clopath synapse.
 
-So far the implementation of the urbanczik_synapse only supports two-compartment neurons.
+Transmits
++++++++++
 
-Transmits: SpikeEvent
+SpikeEvent
 
-References:
+References
+++++++++++
 
-\verbatim embed:rst
-.. [1] R. Urbanczik, W. Senn (2014). Learning by the Dendritic Prediction of Somatic Spiking. Neuron, 81, 521 - 528.
-\endverbatim
-Authors: Jonas Stapmanns, David Dahmen, Jan Hahne
+.. [1] Clopath et al. (2010). Connectivity reflects coding:
+       a model of voltage-based STDP with homeostasis.
+       Nature Neuroscience 13:3, 344--352. DOI: https://doi.org/10.1038/nn.2479
+.. [2] Clopath and Gerstner (2010). Voltage and spike timing interact
+       in STDP – a unified model. Frontiers in Synaptic Neuroscience 2:25.
+       DOI: https://doi.org/10.3389/fnsyn.2010.00025
+.. [3] Voltage-based STDP synapse (Clopath et al. 2010) on ModelDB
+       https://senselab.med.yale.edu/ModelDB/showmodel.cshtml?model=144566
 
-SeeAlso: stdp_synapse, clopath_synapse, pp_cond_exp_mc_urbanczik
-*/
+See also
+++++++++
+
+stdp_synapse, aeif_psc_delta_clopath, hh_psc_alpha_clopath
+
+EndUserDocs */
+
 // connections are templates of target identifier type (used for pointer /
 // target index addressing) derived from generic connection template
 template < typename targetidentifierT >
-class UrbanczikConnection : public Connection< targetidentifierT >
+class clopath_synapse : public Connection< targetidentifierT >
 {
 
 public:
@@ -97,14 +111,14 @@ public:
    * Default Constructor.
    * Sets default values for all parameters. Needed by GenericConnectorModel.
    */
-  UrbanczikConnection();
+  clopath_synapse();
 
 
   /**
    * Copy constructor.
    * Needs to be defined properly in order for GenericConnector to work.
    */
-  UrbanczikConnection( const UrbanczikConnection& ) = default;
+  clopath_synapse( const clopath_synapse& ) = default;
 
   // Explicitly declare all methods inherited from the dependent base
   // ConnectionBase. This avoids explicit name prefixes in all places these
@@ -163,17 +177,26 @@ public:
   }
 
 private:
+  double
+  depress_( double w, double dw )
+  {
+    w -= dw;
+    return w > Wmin_ ? w : Wmin_;
+  }
+
+  double
+  facilitate_( double w, double dw, double x_bar )
+  {
+    w += dw * x_bar;
+    return w < Wmax_ ? w : Wmax_;
+  }
+
   // data members of each connection
   double weight_;
-  double init_weight_;
-  double tau_Delta_;
-  double eta_;
+  double x_bar_;
+  double tau_x_;
   double Wmin_;
   double Wmax_;
-  double PI_integral_;
-  double PI_exp_integral_;
-  double tau_L_trace_;
-  double tau_s_trace_;
 
   double t_lastspike_;
 };
@@ -187,10 +210,11 @@ private:
  */
 template < typename targetidentifierT >
 inline void
-UrbanczikConnection< targetidentifierT >::send( Event& e, thread t, const CommonSynapseProperties& )
+clopath_synapse< targetidentifierT >::send( Event& e, thread t, const CommonSynapseProperties& )
 {
   double t_spike = e.get_stamp().get_ms();
-  // use accessor functions (inherited from Connection< >) to obtain delay and target
+  // use accessor functions (inherited from Connection< >) to obtain delay and
+  // target
   Node* target = get_target( t );
   double dendritic_delay = get_delay();
 
@@ -198,83 +222,61 @@ UrbanczikConnection< targetidentifierT >::send( Event& e, thread t, const Common
   std::deque< histentry_extended >::iterator start;
   std::deque< histentry_extended >::iterator finish;
 
-  // for now we only support two-compartment neurons
-  // in this case the dendritic compartment has index 1
-  const int comp = 1;
-
-  target->get_urbanczik_history( t_lastspike_ - dendritic_delay, t_spike - dendritic_delay, &start, &finish, comp );
-
-  double const g_L = target->get_g_L( comp );
-  double const tau_L = target->get_tau_L( comp );
-  double const C_m = target->get_C_m( comp );
-  double const tau_s = weight_ > 0.0 ? target->get_tau_syn_ex( comp ) : target->get_tau_syn_in( comp );
-  double dPI_exp_integral = 0.0;
-
+  // For a new synapse, t_lastspike_ contains the point in time of the last
+  // spike. So we initially read the
+  // history(t_last_spike - dendritic_delay, ..., T_spike-dendritic_delay]
+  // which increases the access counter for these entries.
+  // At registration, all entries' access counters of
+  // history[0, ..., t_last_spike - dendritic_delay] have been
+  // incremented by ArchivingNode::register_stdp_connection(). See bug #218 for
+  // details.
+  target->get_LTP_history( t_lastspike_ - dendritic_delay, t_spike - dendritic_delay, &start, &finish );
+  // facilitation due to postsynaptic activity since last pre-synaptic spike
   while ( start != finish )
   {
-    double const t_up = start->t_ + dendritic_delay;     // from t_lastspike to t_spike
-    double const minus_delta_t_up = t_lastspike_ - t_up; // from 0 to -delta t
-    double const minus_t_down = t_up - t_spike;          // from -t_spike to 0
-    double const PI =
-      ( tau_L_trace_ * exp( minus_delta_t_up / tau_L ) - tau_s_trace_ * exp( minus_delta_t_up / tau_s ) ) * start->dw_;
-    PI_integral_ += PI;
-    dPI_exp_integral += exp( minus_t_down / tau_Delta_ ) * PI;
-    start++;
+    const double minus_dt = t_lastspike_ - ( start->t_ + dendritic_delay );
+    weight_ = facilitate_( weight_, start->dw_, x_bar_ * exp( minus_dt / tau_x_ ) );
+    ++start;
   }
 
-  PI_exp_integral_ = ( exp( ( t_lastspike_ - t_spike ) / tau_Delta_ ) * PI_exp_integral_ + dPI_exp_integral );
-  weight_ = PI_integral_ - PI_exp_integral_;
-  weight_ = init_weight_ + weight_ * 15.0 * C_m * tau_s * eta_ / ( g_L * ( tau_L - tau_s ) );
-
-  if ( weight_ > Wmax_ )
-  {
-    weight_ = Wmax_;
-  }
-  else if ( weight_ < Wmin_ )
-  {
-    weight_ = Wmin_;
-  }
+  // depression due to new pre-synaptic spike
+  weight_ = depress_( weight_, target->get_LTD_value( t_spike - dendritic_delay ) );
 
   e.set_receiver( *target );
   e.set_weight( weight_ );
-  // use accessor functions (inherited from Connection< >) to obtain delay in steps and rport
+  // use accessor functions (inherited from Connection< >) to obtain delay in
+  // steps and rport
   e.set_delay_steps( get_delay_steps() );
   e.set_rport( get_rport() );
   e();
 
   // compute the trace of the presynaptic spike train
-  tau_L_trace_ = tau_L_trace_ * std::exp( ( t_lastspike_ - t_spike ) / tau_L ) + 1.0;
-  tau_s_trace_ = tau_s_trace_ * std::exp( ( t_lastspike_ - t_spike ) / tau_s ) + 1.0;
+  x_bar_ = x_bar_ * std::exp( ( t_lastspike_ - t_spike ) / tau_x_ ) + 1.0 / tau_x_;
 
   t_lastspike_ = t_spike;
 }
 
 
 template < typename targetidentifierT >
-UrbanczikConnection< targetidentifierT >::UrbanczikConnection()
+clopath_synapse< targetidentifierT >::clopath_synapse()
   : ConnectionBase()
   , weight_( 1.0 )
-  , init_weight_( 1.0 )
-  , tau_Delta_( 100.0 )
-  , eta_( 0.07 )
+  , x_bar_( 0.0 )
+  , tau_x_( 15.0 )
   , Wmin_( 0.0 )
   , Wmax_( 100.0 )
-  , PI_integral_( 0.0 )
-  , PI_exp_integral_( 0.0 )
-  , tau_L_trace_( 0.0 )
-  , tau_s_trace_( 0.0 )
-  , t_lastspike_( -1.0 )
+  , t_lastspike_( 0.0 )
 {
 }
 
 template < typename targetidentifierT >
 void
-UrbanczikConnection< targetidentifierT >::get_status( DictionaryDatum& d ) const
+clopath_synapse< targetidentifierT >::get_status( DictionaryDatum& d ) const
 {
   ConnectionBase::get_status( d );
   def< double >( d, names::weight, weight_ );
-  def< double >( d, names::tau_Delta, tau_Delta_ );
-  def< double >( d, names::eta, eta_ );
+  def< double >( d, names::x_bar, x_bar_ );
+  def< double >( d, names::tau_x, tau_x_ );
   def< double >( d, names::Wmin, Wmin_ );
   def< double >( d, names::Wmax, Wmax_ );
   def< long >( d, names::size_of, sizeof( *this ) );
@@ -282,16 +284,15 @@ UrbanczikConnection< targetidentifierT >::get_status( DictionaryDatum& d ) const
 
 template < typename targetidentifierT >
 void
-UrbanczikConnection< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
+clopath_synapse< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
 {
   ConnectionBase::set_status( d, cm );
   updateValue< double >( d, names::weight, weight_ );
-  updateValue< double >( d, names::tau_Delta, tau_Delta_ );
-  updateValue< double >( d, names::eta, eta_ );
+  updateValue< double >( d, names::x_bar, x_bar_ );
+  updateValue< double >( d, names::tau_x, tau_x_ );
   updateValue< double >( d, names::Wmin, Wmin_ );
   updateValue< double >( d, names::Wmax, Wmax_ );
 
-  init_weight_ = weight_;
   // check if weight_ and Wmin_ has the same sign
   if ( not( ( ( weight_ >= 0 ) - ( weight_ < 0 ) ) == ( ( Wmin_ >= 0 ) - ( Wmin_ < 0 ) ) ) )
   {
@@ -307,4 +308,4 @@ UrbanczikConnection< targetidentifierT >::set_status( const DictionaryDatum& d, 
 
 } // of namespace nest
 
-#endif // of #ifndef URBANCZIK_CONNECTION_H
+#endif // of #ifndef CLOPATH_SYNAPSE_H
