--- models/stdp_nn_pre-centered_connection.h
@@ -1,353 +1,0 @@
-/*
- *  stdp_nn_pre-centered_connection.h
- *
- *  This file is part of NEST.
- *
- *  Copyright (C) 2004 The NEST Initiative
- *
- *  NEST is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 2 of the License, or
- *  (at your option) any later version.
- *
- *  NEST is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with NEST.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-#ifndef STDP_NN_PRE_CENTERED_CONNECTION_H
-#define STDP_NN_PRE_CENTERED_CONNECTION_H
-
-// C++ includes:
-#include <cmath>
-
-// Includes from nestkernel:
-#include "common_synapse_properties.h"
-#include "connection.h"
-#include "connector_model.h"
-#include "event.h"
-
-// Includes from sli:
-#include "dictdatum.h"
-#include "dictutils.h"
-
-namespace nest
-{
-
-/* BeginUserDocs: synapse, spike-timing-dependent plasticity
-
-Short description
-+++++++++++++++++
-
-Synapse type for spike-timing dependent plasticity with
-presynaptic-centered nearest-neighbour spike pairing scheme
-
-Description
-+++++++++++
-
-stdp_nn_pre-centered_synapse is a connector to create synapses with spike
-time dependent plasticity with the presynaptic-centered nearest-neighbour
-spike pairing scheme, as described in [1]_.
-
-Each presynaptic spike is taken into account in the STDP weight change rule
-with the nearest preceding postsynaptic one and the nearest succeeding
-postsynaptic one (instead of pairing with all spikes, like in stdp_synapse).
-So, when a presynaptic spike occurs, it is accounted in the depression rule
-with the nearest preceding postsynaptic one; and when a postsynaptic spike
-occurs, it is accounted in the facilitation rule with all preceding
-presynaptic spikes that were not earlier than the previous postsynaptic
-spike. For a clear illustration of this scheme see fig. 7B in [2]_.
-
-The pairs exactly coinciding (so that presynaptic_spike == postsynaptic_spike
-+ dendritic_delay), leading to zero delta_t, are discarded. In this case the
-concerned pre/postsynaptic spike is paired with the second latest preceding
-post/presynaptic one (for example, pre=={10 ms; 20 ms} and post=={20 ms} will
-result in a potentiation pair 20-to-10).
-
-The implementation involves two additional variables - presynaptic and
-postsynaptic traces [2]_. The presynaptic trace decays exponentially over
-time with the time constant tau_plus, increases by 1 on a pre-spike
-occurrence, and is reset to 0 on a post-spike occurrence. The postsynaptic
-trace (implemented on the postsynaptic neuron side) decays with the time
-constant tau_minus and increases to 1 on a post-spike occurrence.
-
-Parameters
-++++++++++
-
-========= =======  ======================================================
- tau_plus  ms      Time constant of STDP window, potentiation
-                   (tau_minus defined in postsynaptic neuron)
- lambda    real    Step size
- alpha     real    Asymmetry parameter (scales depressing increments as
-                   alpha*lambda)
- mu_plus   real    Weight dependence exponent, potentiation
- mu_minus  real    Weight dependence exponent, depression
- Wmax      real    Maximum allowed weight
-========= =======  ======================================================
-
-Transmits
-+++++++++
-
-SpikeEvent
-
-References
-++++++++++
-
-.. [1] Izhikevich E. M., Desai N. S. (2003) Relating STDP to BCM,
-       Neural Comput. 15, 1511--1523
-
-.. [2] Morrison A., Diesmann M., and Gerstner W. (2008) Phenomenological
-       models of synaptic plasticity based on spike timing,
-       Biol. Cybern. 98, 459--478
-
-See also
-++++++++
-
-stdp_synapse, stdp_nn_symm_synapse
-
-EndUserDocs */
-
-// connections are templates of target identifier type (used for pointer /
-// target index addressing) derived from generic connection template
-
-template < typename targetidentifierT >
-class STDPNNPreCenteredConnection : public Connection< targetidentifierT >
-{
-
-public:
-  typedef CommonSynapseProperties CommonPropertiesType;
-  typedef Connection< targetidentifierT > ConnectionBase;
-
-  /**
-   * Default Constructor.
-   * Sets default values for all parameters. Needed by GenericConnectorModel.
-   */
-  STDPNNPreCenteredConnection();
-
-
-  /**
-   * Copy constructor.
-   * Needs to be defined properly in order for GenericConnector to work.
-   */
-  STDPNNPreCenteredConnection( const STDPNNPreCenteredConnection& ) = default;
-
-  // Explicitly declare all methods inherited from the dependent base
-  // ConnectionBase. This avoids explicit name prefixes in all places these
-  // functions are used. Since ConnectionBase depends on the template parameter,
-  // they are not automatically found in the base class.
-  using ConnectionBase::get_delay_steps;
-  using ConnectionBase::get_delay;
-  using ConnectionBase::get_rport;
-  using ConnectionBase::get_target;
-
-  /**
-   * Get all properties of this connection and put them into a dictionary.
-   */
-  void get_status( DictionaryDatum& d ) const;
-
-  /**
-   * Set properties of this connection from the values given in dictionary.
-   */
-  void set_status( const DictionaryDatum& d, ConnectorModel& cm );
-
-  /**
-   * Send an event to the receiver of this connection.
-   * \param e The event to send
-   * \param cp common properties of all synapses (empty).
-   */
-  void send( Event& e, thread t, const CommonSynapseProperties& cp );
-
-
-  class ConnTestDummyNode : public ConnTestDummyNodeBase
-  {
-  public:
-    // Ensure proper overriding of overloaded virtual functions.
-    // Return values from functions are ignored.
-    using ConnTestDummyNodeBase::handles_test_event;
-    port
-    handles_test_event( SpikeEvent&, rport )
-    {
-      return invalid_port_;
-    }
-  };
-
-  void
-  check_connection( Node& s, Node& t, rport receptor_type, const CommonPropertiesType& )
-  {
-    ConnTestDummyNode dummy_target;
-
-    ConnectionBase::check_connection_( dummy_target, s, t, receptor_type );
-
-    t.register_stdp_connection( t_lastspike_ - get_delay(), get_delay() );
-  }
-
-  void
-  set_weight( double w )
-  {
-    weight_ = w;
-  }
-
-private:
-  double
-  facilitate_( double w, double kplus )
-  {
-    double norm_w = ( w / Wmax_ ) + ( lambda_ * std::pow( 1.0 - ( w / Wmax_ ), mu_plus_ ) * kplus );
-    return norm_w < 1.0 ? norm_w * Wmax_ : Wmax_;
-  }
-
-  double
-  depress_( double w, double kminus )
-  {
-    double norm_w = ( w / Wmax_ ) - ( alpha_ * lambda_ * std::pow( w / Wmax_, mu_minus_ ) * kminus );
-    return norm_w > 0.0 ? norm_w * Wmax_ : 0.0;
-  }
-
-  // data members of each connection
-  double weight_;
-  double tau_plus_;
-  double lambda_;
-  double alpha_;
-  double mu_plus_;
-  double mu_minus_;
-  double Wmax_;
-  double Kplus_;
-
-  double t_lastspike_;
-};
-
-
-/**
- * Send an event to the receiver of this connection.
- * \param e The event to send
- * \param t The thread on which this connection is stored.
- * \param cp Common properties object, containing the stdp parameters.
- */
-template < typename targetidentifierT >
-inline void
-STDPNNPreCenteredConnection< targetidentifierT >::send( Event& e, thread t, const CommonSynapseProperties& )
-{
-  // synapse STDP depressing/facilitation dynamics
-  double t_spike = e.get_stamp().get_ms();
-
-  // use accessor functions (inherited from Connection< >) to obtain delay and
-  // target
-  Node* target = get_target( t );
-  double dendritic_delay = get_delay();
-
-  // get spike history in relevant range (t1, t2] from postsynaptic neuron
-  std::deque< histentry >::iterator start;
-  std::deque< histentry >::iterator finish;
-
-  // For a new synapse, t_lastspike_ contains the point in time of the last
-  // spike. So we initially read the
-  // history(t_last_spike - dendritic_delay, ..., T_spike-dendritic_delay]
-  // which increases the access counter for these entries.
-  // At registration, all entries' access counters of
-  // history[0, ..., t_last_spike - dendritic_delay] have been
-  // incremented by ArchivingNode::register_stdp_connection(). See bug #218 for
-  // details.
-  target->get_history( t_lastspike_ - dendritic_delay, t_spike - dendritic_delay, &start, &finish );
-  // If there were no postsynaptic spikes between the current pre-synaptic one
-  // t_spike and the previous pre-synaptic one t_lastspike_, there are no pairs
-  // to account.
-  if ( start != finish )
-  {
-    // facilitation due to the first postsynaptic spike start->t_
-    // since the previous pre-synaptic spike t_lastspike_
-
-    double minus_dt;
-    minus_dt = t_lastspike_ - ( start->t_ + dendritic_delay );
-
-    // get_history() should make sure that
-    // start->t_ > t_lastspike_ - dendritic_delay, i.e. minus_dt < 0
-    assert( minus_dt < -1.0 * kernel().connection_manager.get_stdp_eps() );
-
-    weight_ = facilitate_( weight_, Kplus_ * std::exp( minus_dt / tau_plus_ ) );
-
-    // According to the presynaptic-centered nearest-neighbour scheme,
-    // a postsynaptic spike
-    // (we now know there was at least one between t_lastspike_ and t_spike)
-    // erases the state of the synapse,
-    // and all the preceding presynaptic spikes are forgotten.
-    Kplus_ = 0;
-  }
-
-  // depression due to the latest postsynaptic spike finish->t_
-  // before the current pre-synaptic spike t_spike
-  double nearest_neighbor_Kminus;
-  double value_to_throw_away; // discard Kminus and Kminus_triplet here
-  target->get_K_values( t_spike - dendritic_delay, value_to_throw_away, nearest_neighbor_Kminus, value_to_throw_away );
-  weight_ = depress_( weight_, nearest_neighbor_Kminus );
-
-  Kplus_ = Kplus_ * std::exp( ( t_lastspike_ - t_spike ) / tau_plus_ ) + 1.0;
-
-  e.set_receiver( *target );
-  e.set_weight( weight_ );
-  // use accessor functions (inherited from Connection< >) to obtain delay in
-  // steps and rport
-  e.set_delay_steps( get_delay_steps() );
-  e.set_rport( get_rport() );
-  e();
-
-  t_lastspike_ = t_spike;
-}
-
-
-template < typename targetidentifierT >
-STDPNNPreCenteredConnection< targetidentifierT >::STDPNNPreCenteredConnection()
-  : ConnectionBase()
-  , weight_( 1.0 )
-  , tau_plus_( 20.0 )
-  , lambda_( 0.01 )
-  , alpha_( 1.0 )
-  , mu_plus_( 1.0 )
-  , mu_minus_( 1.0 )
-  , Wmax_( 100.0 )
-  , Kplus_( 0.0 )
-  , t_lastspike_( 0.0 )
-{
-}
-
-template < typename targetidentifierT >
-void
-STDPNNPreCenteredConnection< targetidentifierT >::get_status( DictionaryDatum& d ) const
-{
-  ConnectionBase::get_status( d );
-  def< double >( d, names::weight, weight_ );
-  def< double >( d, names::tau_plus, tau_plus_ );
-  def< double >( d, names::lambda, lambda_ );
-  def< double >( d, names::alpha, alpha_ );
-  def< double >( d, names::mu_plus, mu_plus_ );
-  def< double >( d, names::mu_minus, mu_minus_ );
-  def< double >( d, names::Wmax, Wmax_ );
-  def< long >( d, names::size_of, sizeof( *this ) );
-}
-
-template < typename targetidentifierT >
-void
-STDPNNPreCenteredConnection< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
-{
-  ConnectionBase::set_status( d, cm );
-  updateValue< double >( d, names::weight, weight_ );
-  updateValue< double >( d, names::tau_plus, tau_plus_ );
-  updateValue< double >( d, names::lambda, lambda_ );
-  updateValue< double >( d, names::alpha, alpha_ );
-  updateValue< double >( d, names::mu_plus, mu_plus_ );
-  updateValue< double >( d, names::mu_minus, mu_minus_ );
-  updateValue< double >( d, names::Wmax, Wmax_ );
-
-  // check if weight_ and Wmax_ have the same sign
-  if ( not( ( ( weight_ >= 0 ) - ( weight_ < 0 ) ) == ( ( Wmax_ >= 0 ) - ( Wmax_ < 0 ) ) ) )
-  {
-    throw BadProperty( "Weight and Wmax must have same sign." );
-  }
-}
-
-} // of namespace nest
-
-#endif // of #ifndef STDP_NN_PRE_CENTERED_CONNECTION_H
