+++ models/stdp_nn_symm_synapse.h
@@ -1,5 +1,5 @@
 /*
- *  clopath_connection.h
+ *  stdp_nn_symm_synapse.h
  *
  *  This file is part of NEST.
  *
@@ -20,8 +20,8 @@
  *
  */
 
-#ifndef CLOPATH_CONNECTION_H
-#define CLOPATH_CONNECTION_H
+#ifndef STDP_NN_SYMM_SYNAPSE_H
+#define STDP_NN_SYMM_SYNAPSE_H
 
 // C++ includes:
 #include <cmath>
@@ -31,7 +31,6 @@
 #include "connection.h"
 #include "connector_model.h"
 #include "event.h"
-#include "ring_buffer.h"
 
 // Includes from sli:
 #include "dictdatum.h"
@@ -40,38 +39,53 @@
 namespace nest
 {
 
-/* BeginUserDocs: synapse, spike-timing-dependent plasticity, Clopath plasticity
+/* BeginUserDocs: synapse, spike-timing-dependent plasticity
 
 Short description
 +++++++++++++++++
 
-Synapse type for voltage-based STDP after Clopath
+Synapse type for spike-timing dependent plasticity with symmetric nearest-neighbour spike pairing scheme
 
 Description
 +++++++++++
 
-clopath_synapse is a connector to create Clopath synapses as defined
-in [1]_. In contrast to usual STDP, the change of the synaptic weight does
-not only depend on the pre- and postsynaptic spike timing but also on the
-postsynaptic membrane potential.
+stdp_nn_symm_synapse is a connector to create synapses with spike time
+dependent plasticity with the symmetric nearest-neighbour spike pairing
+scheme [1]_.
 
-Clopath synapses require archiving of continuous quantities. Therefore Clopath
-synapses can only be connected to neuron models that are capable of doing this
-archiving. So far, compatible models are aeif_psc_delta_clopath and
-hh_psc_alpha_clopath.
+When a presynaptic spike occurs, it is taken into account in the depression
+part of the STDP weight change rule with the nearest preceding postsynaptic
+one, and when a postsynaptic spike occurs, it is accounted in the
+facilitation rule with the nearest preceding presynaptic one (instead of
+pairing with all spikes, like in stdp_synapse). For a clear illustration of
+this scheme see fig. 7A in [2]_.
+
+The pairs exactly coinciding (so that presynaptic_spike == postsynaptic_spike
++ dendritic_delay), leading to zero delta_t, are discarded. In this case the
+concerned pre/postsynaptic spike is paired with the second latest preceding
+post/presynaptic one (for example, pre=={10 ms; 20 ms} and post=={20 ms} will
+result in a potentiation pair 20-to-10).
+
+The implementation involves two additional variables - presynaptic and
+postsynaptic traces [2]_. The presynaptic trace decays exponentially over
+time with the time constant tau_plus and increases to 1 on a pre-spike
+occurrence. The postsynaptic trace (implemented on the postsynaptic neuron
+side) decays with the time constant tau_minus and increases to 1 on a
+post-spike occurrence.
 
 Parameters
 ++++++++++
 
-=======  ======  ==========================================================
-tau_x    ms      Time constant of the trace of the presynaptic spike train
-Wmax     real    Maximum allowed weight
-Wmin     real    Minimum allowed weight
-=======  ======  ==========================================================
-
-Other parameters like the amplitudes for long-term potentiation (LTP) and
-depression (LTD) are stored in in the neuron models that are compatible with the
-Clopath synapse.
+========= =======  ======================================================
+ tau_plus  ms      Time constant of STDP window, potentiation
+                   (tau_minus defined in postsynaptic neuron)
+ lambda    real    Step size
+ alpha     real    Asymmetry parameter (scales depressing increments as
+                   alpha*lambda)
+ mu_plus   real    Weight dependence exponent, potentiation
+ mu_minus  real    Weight dependence exponent, depression
+ Wmax      real    Maximum allowed weight
+========= =======  ======================================================
 
 Transmits
 +++++++++
@@ -81,26 +95,25 @@ SpikeEvent
 References
 ++++++++++
 
-.. [1] Clopath et al. (2010). Connectivity reflects coding:
-       a model of voltage-based STDP with homeostasis.
-       Nature Neuroscience 13:3, 344--352. DOI: https://doi.org/10.1038/nn.2479
-.. [2] Clopath and Gerstner (2010). Voltage and spike timing interact
-       in STDP – a unified model. Frontiers in Synaptic Neuroscience 2:25.
-       DOI: https://doi.org/10.3389/fnsyn.2010.00025
-.. [3] Voltage-based STDP synapse (Clopath et al. 2010) on ModelDB
-       https://senselab.med.yale.edu/ModelDB/showmodel.cshtml?model=144566
+.. [1] Morrison A., Aertsen A., Diesmann M. (2007) Spike-timing dependent
+       plasticity in balanced random networks, Neural Comput. 19:1437--1467
+
+.. [2] Morrison A., Diesmann M., and Gerstner W. (2008) Phenomenological
+       models of synaptic plasticity based on spike timing,
+       Biol. Cybern. 98, 459--478
 
 See also
 ++++++++
 
-stdp_synapse, aeif_psc_delta_clopath, hh_psc_alpha_clopath
+stdp_synapse
 
 EndUserDocs */
 
 // connections are templates of target identifier type (used for pointer /
 // target index addressing) derived from generic connection template
+
 template < typename targetidentifierT >
-class ClopathConnection : public Connection< targetidentifierT >
+class stdp_nn_symm_synapse : public Connection< targetidentifierT >
 {
 
 public:
@@ -111,14 +124,14 @@ public:
    * Default Constructor.
    * Sets default values for all parameters. Needed by GenericConnectorModel.
    */
-  ClopathConnection();
+  stdp_nn_symm_synapse();
 
 
   /**
    * Copy constructor.
    * Needs to be defined properly in order for GenericConnector to work.
    */
-  ClopathConnection( const ClopathConnection& ) = default;
+  stdp_nn_symm_synapse( const stdp_nn_symm_synapse& ) = default;
 
   // Explicitly declare all methods inherited from the dependent base
   // ConnectionBase. This avoids explicit name prefixes in all places these
@@ -178,24 +191,26 @@ public:
 
 private:
   double
-  depress_( double w, double dw )
+  facilitate_( double w, double kplus )
   {
-    w -= dw;
-    return w > Wmin_ ? w : Wmin_;
+    double norm_w = ( w / Wmax_ ) + ( lambda_ * std::pow( 1.0 - ( w / Wmax_ ), mu_plus_ ) * kplus );
+    return norm_w < 1.0 ? norm_w * Wmax_ : Wmax_;
   }
 
   double
-  facilitate_( double w, double dw, double x_bar )
+  depress_( double w, double kminus )
   {
-    w += dw * x_bar;
-    return w < Wmax_ ? w : Wmax_;
+    double norm_w = ( w / Wmax_ ) - ( alpha_ * lambda_ * std::pow( w / Wmax_, mu_minus_ ) * kminus );
+    return norm_w > 0.0 ? norm_w * Wmax_ : 0.0;
   }
 
   // data members of each connection
   double weight_;
-  double x_bar_;
-  double tau_x_;
-  double Wmin_;
+  double tau_plus_;
+  double lambda_;
+  double alpha_;
+  double mu_plus_;
+  double mu_minus_;
   double Wmax_;
 
   double t_lastspike_;
@@ -206,21 +221,23 @@ private:
  * Send an event to the receiver of this connection.
  * \param e The event to send
  * \param t The thread on which this connection is stored.
- * \param cp Common properties object, containing the stdp parameters.
+  * \param cp Common properties object, containing the stdp parameters.
  */
 template < typename targetidentifierT >
 inline void
-ClopathConnection< targetidentifierT >::send( Event& e, thread t, const CommonSynapseProperties& )
+stdp_nn_symm_synapse< targetidentifierT >::send( Event& e, thread t, const CommonSynapseProperties& )
 {
+  // synapse STDP depressing/facilitation dynamics
   double t_spike = e.get_stamp().get_ms();
+
   // use accessor functions (inherited from Connection< >) to obtain delay and
   // target
   Node* target = get_target( t );
   double dendritic_delay = get_delay();
 
   // get spike history in relevant range (t1, t2] from postsynaptic neuron
-  std::deque< histentry_extended >::iterator start;
-  std::deque< histentry_extended >::iterator finish;
+  std::deque< histentry >::iterator start;
+  std::deque< histentry >::iterator finish;
 
   // For a new synapse, t_lastspike_ contains the point in time of the last
   // spike. So we initially read the
@@ -230,17 +247,26 @@ ClopathConnection< targetidentifierT >::send( Event& e, thread t, const CommonSy
   // history[0, ..., t_last_spike - dendritic_delay] have been
   // incremented by ArchivingNode::register_stdp_connection(). See bug #218 for
   // details.
-  target->get_LTP_history( t_lastspike_ - dendritic_delay, t_spike - dendritic_delay, &start, &finish );
-  // facilitation due to postsynaptic activity since last pre-synaptic spike
+  target->get_history( t_lastspike_ - dendritic_delay, t_spike - dendritic_delay, &start, &finish );
+  // facilitation due to postsynaptic spikes since the last pre-synaptic spike
+  double minus_dt;
   while ( start != finish )
   {
-    const double minus_dt = t_lastspike_ - ( start->t_ + dendritic_delay );
-    weight_ = facilitate_( weight_, start->dw_, x_bar_ * exp( minus_dt / tau_x_ ) );
+    minus_dt = t_lastspike_ - ( start->t_ + dendritic_delay );
     ++start;
+
+    // get_history() should make sure that
+    // start->t_ > t_lastspike_ - dendritic_delay, i.e. minus_dt < 0
+    assert( minus_dt < -1.0 * kernel().connection_manager.get_stdp_eps() );
+
+    weight_ = facilitate_( weight_, std::exp( minus_dt / tau_plus_ ) );
   }
 
-  // depression due to new pre-synaptic spike
-  weight_ = depress_( weight_, target->get_LTD_value( t_spike - dendritic_delay ) );
+  // depression due to the new pre-synaptic spike
+  double nearest_neighbor_Kminus;
+  double value_to_throw_away; // discard Kminus and Kminus_triplet here
+  target->get_K_values( t_spike - dendritic_delay, value_to_throw_away, nearest_neighbor_Kminus, value_to_throw_away );
+  weight_ = depress_( weight_, nearest_neighbor_Kminus );
 
   e.set_receiver( *target );
   e.set_weight( weight_ );
@@ -250,20 +276,19 @@ ClopathConnection< targetidentifierT >::send( Event& e, thread t, const CommonSy
   e.set_rport( get_rport() );
   e();
 
-  // compute the trace of the presynaptic spike train
-  x_bar_ = x_bar_ * std::exp( ( t_lastspike_ - t_spike ) / tau_x_ ) + 1.0 / tau_x_;
-
   t_lastspike_ = t_spike;
 }
 
 
 template < typename targetidentifierT >
-ClopathConnection< targetidentifierT >::ClopathConnection()
+stdp_nn_symm_synapse< targetidentifierT >::stdp_nn_symm_synapse()
   : ConnectionBase()
   , weight_( 1.0 )
-  , x_bar_( 0.0 )
-  , tau_x_( 15.0 )
-  , Wmin_( 0.0 )
+  , tau_plus_( 20.0 )
+  , lambda_( 0.01 )
+  , alpha_( 1.0 )
+  , mu_plus_( 1.0 )
+  , mu_minus_( 1.0 )
   , Wmax_( 100.0 )
   , t_lastspike_( 0.0 )
 {
@@ -271,36 +296,34 @@ ClopathConnection< targetidentifierT >::ClopathConnection()
 
 template < typename targetidentifierT >
 void
-ClopathConnection< targetidentifierT >::get_status( DictionaryDatum& d ) const
+stdp_nn_symm_synapse< targetidentifierT >::get_status( DictionaryDatum& d ) const
 {
   ConnectionBase::get_status( d );
   def< double >( d, names::weight, weight_ );
-  def< double >( d, names::x_bar, x_bar_ );
-  def< double >( d, names::tau_x, tau_x_ );
-  def< double >( d, names::Wmin, Wmin_ );
+  def< double >( d, names::tau_plus, tau_plus_ );
+  def< double >( d, names::lambda, lambda_ );
+  def< double >( d, names::alpha, alpha_ );
+  def< double >( d, names::mu_plus, mu_plus_ );
+  def< double >( d, names::mu_minus, mu_minus_ );
   def< double >( d, names::Wmax, Wmax_ );
   def< long >( d, names::size_of, sizeof( *this ) );
 }
 
 template < typename targetidentifierT >
 void
-ClopathConnection< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
+stdp_nn_symm_synapse< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
 {
   ConnectionBase::set_status( d, cm );
   updateValue< double >( d, names::weight, weight_ );
-  updateValue< double >( d, names::x_bar, x_bar_ );
-  updateValue< double >( d, names::tau_x, tau_x_ );
-  updateValue< double >( d, names::Wmin, Wmin_ );
+  updateValue< double >( d, names::tau_plus, tau_plus_ );
+  updateValue< double >( d, names::lambda, lambda_ );
+  updateValue< double >( d, names::alpha, alpha_ );
+  updateValue< double >( d, names::mu_plus, mu_plus_ );
+  updateValue< double >( d, names::mu_minus, mu_minus_ );
   updateValue< double >( d, names::Wmax, Wmax_ );
 
-  // check if weight_ and Wmin_ has the same sign
-  if ( not( ( ( weight_ >= 0 ) - ( weight_ < 0 ) ) == ( ( Wmin_ >= 0 ) - ( Wmin_ < 0 ) ) ) )
-  {
-    throw BadProperty( "Weight and Wmin must have same sign." );
-  }
-
-  // check if weight_ and Wmax_ has the same sign
-  if ( not( ( ( weight_ >= 0 ) - ( weight_ < 0 ) ) == ( ( Wmax_ > 0 ) - ( Wmax_ <= 0 ) ) ) )
+  // check if weight_ and Wmax_ have the same sign
+  if ( not( ( ( weight_ >= 0 ) - ( weight_ < 0 ) ) == ( ( Wmax_ >= 0 ) - ( Wmax_ < 0 ) ) ) )
   {
     throw BadProperty( "Weight and Wmax must have same sign." );
   }
@@ -308,4 +331,4 @@ ClopathConnection< targetidentifierT >::set_status( const DictionaryDatum& d, Co
 
 } // of namespace nest
 
-#endif // of #ifndef CLOPATH_CONNECTION_H
+#endif // of #ifndef STDP_NN_SYMM_SYNAPSE_H
