+++ models/stdp_pl_synapse_hom.cpp
@@ -1,5 +1,5 @@
 /*
- *  stdp_connection_hom.cpp
+ *  stdp_pl_synapse_hom.cpp
  *
  *  This file is part of NEST.
  *
@@ -20,7 +20,7 @@
  *
  */
 
-#include "stdp_connection_hom.h"
+#include "stdp_pl_synapse_hom.h"
 
 // Includes from nestkernel:
 #include "common_synapse_properties.h"
@@ -30,47 +30,52 @@
 // Includes from sli:
 #include "dictdatum.h"
 
+
 namespace nest
 {
+
 //
-// Implementation of class STDPHomCommonProperties.
+// Implementation of class STDPPLHomCommonProperties.
 //
 
-STDPHomCommonProperties::STDPHomCommonProperties()
+STDPPLHomCommonProperties::STDPPLHomCommonProperties()
   : CommonSynapseProperties()
   , tau_plus_( 20.0 )
-  , lambda_( 0.01 )
+  , tau_plus_inv_( 1. / tau_plus_ )
+  , lambda_( 0.1 )
   , alpha_( 1.0 )
-  , mu_plus_( 1.0 )
-  , mu_minus_( 1.0 )
-  , Wmax_( 100.0 )
+  , mu_( 0.4 )
 {
 }
 
 void
-STDPHomCommonProperties::get_status( DictionaryDatum& d ) const
+STDPPLHomCommonProperties::get_status( DictionaryDatum& d ) const
 {
   CommonSynapseProperties::get_status( d );
 
   def< double >( d, names::tau_plus, tau_plus_ );
   def< double >( d, names::lambda, lambda_ );
   def< double >( d, names::alpha, alpha_ );
-  def< double >( d, names::mu_plus, mu_plus_ );
-  def< double >( d, names::mu_minus, mu_minus_ );
-  def< double >( d, names::Wmax, Wmax_ );
+  def< double >( d, names::mu, mu_ );
 }
 
 void
-STDPHomCommonProperties::set_status( const DictionaryDatum& d, ConnectorModel& cm )
+STDPPLHomCommonProperties::set_status( const DictionaryDatum& d, ConnectorModel& cm )
 {
   CommonSynapseProperties::set_status( d, cm );
 
   updateValue< double >( d, names::tau_plus, tau_plus_ );
+  if ( tau_plus_ > 0. )
+  {
+    tau_plus_inv_ = 1. / tau_plus_;
+  }
+  else
+  {
+    throw BadProperty( "tau_plus > 0. required." );
+  }
   updateValue< double >( d, names::lambda, lambda_ );
   updateValue< double >( d, names::alpha, alpha_ );
-  updateValue< double >( d, names::mu_plus, mu_plus_ );
-  updateValue< double >( d, names::mu_minus, mu_minus_ );
-  updateValue< double >( d, names::Wmax, Wmax_ );
+  updateValue< double >( d, names::mu, mu_ );
 }
 
 } // of namespace nest
