+++ models/stdp_synapse.h
@@ -1,5 +1,5 @@
 /*
- *  stdp_connection_hom.h
+ *  stdp_synapse.h
  *
  *  This file is part of NEST.
  *
@@ -20,14 +20,21 @@
  *
  */
 
-#ifndef STDP_CONNECTION_HOM_H
-#define STDP_CONNECTION_HOM_H
+#ifndef STDP_SYNAPSE_H
+#define STDP_SYNAPSE_H
 
 // C++ includes:
 #include <cmath>
 
 // Includes from nestkernel:
+#include "common_synapse_properties.h"
 #include "connection.h"
+#include "connector_model.h"
+#include "event.h"
+
+// Includes from sli:
+#include "dictdatum.h"
+#include "dictutils.h"
 
 namespace nest
 {
@@ -37,43 +44,28 @@ namespace nest
 Short description
 +++++++++++++++++
 
-Synapse type for spike-timing dependent plasticity using homogeneous parameters
+Synapse type for spike-timing dependent plasticity
 
 Description
 +++++++++++
 
-stdp_synapse_hom is a connector to create synapses with spike time
+stdp_synapse is a connector to create synapses with spike time
 dependent plasticity (as defined in [1]_). Here the weight dependence
 exponent can be set separately for potentiation and depression.
-
- Parameters controlling plasticity are identical for all synapses of the
- model, reducing the memory required per synapse considerably.
-
-Examples:
-
-* multiplicative STDP [2]_  mu_plus = mu_minus = 1.0
-* additive STDP       [3]_  mu_plus = mu_minus = 0.0
-* Guetig STDP         [1]_  mu_plus = mu_minus = [0.0,1.0]
-* van Rossum STDP     [4]_  mu_plus = 0.0 mu_minus = 1.0
 
 Parameters
 ++++++++++
 
 ========= =======  ======================================================
- tau_plus ms       Time constant of STDP window, potentiation
+ tau_plus  ms      Time constant of STDP window, potentiation
                    (tau_minus defined in postsynaptic neuron)
- lambda   real     Step size
- alpha    real     Asymmetry parameter (scales depressing increments as
+ lambda    real    Step size
+ alpha     real    Asymmetry parameter (scales depressing increments as
                    alpha*lambda)
- mu_plus  real     Weight dependence exponent, potentiation
- mu_minus real     Weight dependence exponent, depression
- Wmax     real     Maximum allowed weight
+ mu_plus   real    Weight dependence exponent, potentiation
+ mu_minus  real    Weight dependence exponent, depression
+ Wmax      real    Maximum allowed weight
 ========= =======  ======================================================
-
-Remarks:
-
-The parameters are common to all synapses of the model and must be set using
-SetDefaults on the synapse model.
 
 Transmits
 +++++++++
@@ -105,72 +97,36 @@ tsodyks_synapse, static_synapse
 
 EndUserDocs */
 
-/**
- * Class containing the common properties for all synapses of type
- * STDPConnectionHom.
- */
+// connections are templates of target identifier type (used for pointer /
+// target index addressing) derived from generic connection template
 
-class STDPHomCommonProperties : public CommonSynapseProperties
-{
-
-public:
-  /**
-   * Default constructor.
-   * Sets all property values to defaults.
-   */
-  STDPHomCommonProperties();
-
-  /**
-   * Get all properties and put them into a dictionary.
-   */
-  void get_status( DictionaryDatum& d ) const;
-
-  /**
-   * Set properties from the values given in dictionary.
-   */
-  void set_status( const DictionaryDatum& d, ConnectorModel& cm );
-
-  // data members common to all connections
-  double tau_plus_;
-  double lambda_;
-  double alpha_;
-  double mu_plus_;
-  double mu_minus_;
-  double Wmax_;
-};
-
-
-/**
- * Class representing an STDP connection with homogeneous parameters, i.e.
- * parameters are the same for all synapses.
- */
 template < typename targetidentifierT >
-class STDPConnectionHom : public Connection< targetidentifierT >
+class stdp_synapse : public Connection< targetidentifierT >
 {
 
 public:
-  typedef STDPHomCommonProperties CommonPropertiesType;
+  typedef CommonSynapseProperties CommonPropertiesType;
   typedef Connection< targetidentifierT > ConnectionBase;
 
   /**
    * Default Constructor.
    * Sets default values for all parameters. Needed by GenericConnectorModel.
    */
-  STDPConnectionHom();
+  stdp_synapse();
+
 
   /**
-   * Copy constructor from a property object.
+   * Copy constructor.
    * Needs to be defined properly in order for GenericConnector to work.
    */
-  STDPConnectionHom( const STDPConnectionHom& ) = default;
-
+  stdp_synapse( const stdp_synapse& ) = default;
 
   // Explicitly declare all methods inherited from the dependent base
   // ConnectionBase. This avoids explicit name prefixes in all places these
   // functions are used. Since ConnectionBase depends on the template parameter,
   // they are not automatically found in the base class.
-  using ConnectionBase::get_delay;
   using ConnectionBase::get_delay_steps;
+  using ConnectionBase::get_delay;
   using ConnectionBase::get_rport;
   using ConnectionBase::get_target;
 
@@ -187,14 +143,9 @@ public:
   /**
    * Send an event to the receiver of this connection.
    * \param e The event to send
+   * \param cp common properties of all synapses (empty).
    */
-  void send( Event& e, thread t, const STDPHomCommonProperties& );
-
-  void
-  set_weight( double w )
-  {
-    weight_ = w;
-  }
+  void send( Event& e, thread t, const CommonSynapseProperties& cp );
 
 
   class ConnTestDummyNode : public ConnTestDummyNodeBase
@@ -210,84 +161,81 @@ public:
     }
   };
 
-  /*
-   * This function calls check_connection on the sender and checks if the
-   * receiver accepts the event type and receptor type requested by the sender.
-   * Node::check_connection() will either confirm the receiver port by returning
-   * true or false if the connection should be ignored.
-   * We have to override the base class' implementation, since for STDP
-   * connections we have to call register_stdp_connection on the target neuron
-   * to inform the Archiver to collect spikes for this connection.
-   *
-   * \param s The source node
-   * \param r The target node
-   * \param receptor_type The ID of the requested receptor type
-   */
   void
   check_connection( Node& s, Node& t, rport receptor_type, const CommonPropertiesType& )
   {
     ConnTestDummyNode dummy_target;
+
     ConnectionBase::check_connection_( dummy_target, s, t, receptor_type );
 
     t.register_stdp_connection( t_lastspike_ - get_delay(), get_delay() );
   }
 
+  void
+  set_weight( double w )
+  {
+    weight_ = w;
+  }
+
 private:
   double
-  facilitate_( double w, double kplus, const STDPHomCommonProperties& cp )
+  facilitate_( double w, double kplus )
   {
-    double norm_w = ( w / cp.Wmax_ ) + ( cp.lambda_ * std::pow( 1.0 - ( w / cp.Wmax_ ), cp.mu_plus_ ) * kplus );
-    return norm_w < 1.0 ? norm_w * cp.Wmax_ : cp.Wmax_;
+    double norm_w = ( w / Wmax_ ) + ( lambda_ * std::pow( 1.0 - ( w / Wmax_ ), mu_plus_ ) * kplus );
+    return norm_w < 1.0 ? norm_w * Wmax_ : Wmax_;
   }
 
   double
-  depress_( double w, double kminus, const STDPHomCommonProperties& cp )
+  depress_( double w, double kminus )
   {
-    double norm_w = ( w / cp.Wmax_ ) - ( cp.alpha_ * cp.lambda_ * std::pow( w / cp.Wmax_, cp.mu_minus_ ) * kminus );
-    return norm_w > 0.0 ? norm_w * cp.Wmax_ : 0.0;
+    double norm_w = ( w / Wmax_ ) - ( alpha_ * lambda_ * std::pow( w / Wmax_, mu_minus_ ) * kminus );
+    return norm_w > 0.0 ? norm_w * Wmax_ : 0.0;
   }
 
   // data members of each connection
   double weight_;
+  double tau_plus_;
+  double lambda_;
+  double alpha_;
+  double mu_plus_;
+  double mu_minus_;
+  double Wmax_;
   double Kplus_;
+
   double t_lastspike_;
 };
 
 
-//
-// Implementation of class STDPConnectionHom.
-//
-
-template < typename targetidentifierT >
-STDPConnectionHom< targetidentifierT >::STDPConnectionHom()
-  : ConnectionBase()
-  , weight_( 1.0 )
-  , Kplus_( 0.0 )
-  , t_lastspike_( 0.0 )
-{
-}
-
 /**
  * Send an event to the receiver of this connection.
  * \param e The event to send
- * \param p The port under which this connection is stored in the Connector.
+ * \param t The thread on which this connection is stored.
+ * \param cp Common properties object, containing the stdp parameters.
  */
 template < typename targetidentifierT >
 inline void
-STDPConnectionHom< targetidentifierT >::send( Event& e, thread t, const STDPHomCommonProperties& cp )
+stdp_synapse< targetidentifierT >::send( Event& e, thread t, const CommonSynapseProperties& )
 {
   // synapse STDP depressing/facilitation dynamics
-
   const double t_spike = e.get_stamp().get_ms();
 
-  // t_lastspike_ = 0 initially
-
+  // use accessor functions (inherited from Connection< >) to obtain delay and
+  // target
   Node* target = get_target( t );
   double dendritic_delay = get_delay();
 
   // get spike history in relevant range (t1, t2] from postsynaptic neuron
   std::deque< histentry >::iterator start;
   std::deque< histentry >::iterator finish;
+
+  // For a new synapse, t_lastspike_ contains the point in time of the last
+  // spike. So we initially read the
+  // history(t_last_spike - dendritic_delay, ..., T_spike-dendritic_delay]
+  // which increases the access counter for these entries.
+  // At registration, all entries' access counters of
+  // history[0, ..., t_last_spike - dendritic_delay] have been
+  // incremented by ArchivingNode::register_stdp_connection(). See bug #218 for
+  // details.
   target->get_history( t_lastspike_ - dendritic_delay, t_spike - dendritic_delay, &start, &finish );
   // facilitation due to postsynaptic spikes since last pre-synaptic spike
   double minus_dt;
@@ -298,48 +246,76 @@ STDPConnectionHom< targetidentifierT >::send( Event& e, thread t, const STDPHomC
     // get_history() should make sure that
     // start->t_ > t_lastspike - dendritic_delay, i.e. minus_dt < 0
     assert( minus_dt < -1.0 * kernel().connection_manager.get_stdp_eps() );
-    weight_ = facilitate_( weight_, Kplus_ * std::exp( minus_dt / cp.tau_plus_ ), cp );
+    weight_ = facilitate_( weight_, Kplus_ * std::exp( minus_dt / tau_plus_ ) );
   }
 
-  // depression due to new pre-synaptic spike
-  weight_ = depress_( weight_, target->get_K_value( t_spike - dendritic_delay ), cp );
+  const double _K_value = target->get_K_value( t_spike - dendritic_delay );
+  weight_ = depress_( weight_, _K_value );
 
   e.set_receiver( *target );
   e.set_weight( weight_ );
+  // use accessor functions (inherited from Connection< >) to obtain delay in
+  // steps and rport
   e.set_delay_steps( get_delay_steps() );
   e.set_rport( get_rport() );
   e();
 
-  Kplus_ = Kplus_ * std::exp( ( t_lastspike_ - t_spike ) / cp.tau_plus_ ) + 1.0;
+  Kplus_ = Kplus_ * std::exp( ( t_lastspike_ - t_spike ) / tau_plus_ ) + 1.0;
 
   t_lastspike_ = t_spike;
 }
 
+
+template < typename targetidentifierT >
+stdp_synapse< targetidentifierT >::stdp_synapse()
+  : ConnectionBase()
+  , weight_( 1.0 )
+  , tau_plus_( 20.0 )
+  , lambda_( 0.01 )
+  , alpha_( 1.0 )
+  , mu_plus_( 1.0 )
+  , mu_minus_( 1.0 )
+  , Wmax_( 100.0 )
+  , Kplus_( 0.0 )
+  , t_lastspike_( 0.0 )
+{
+}
+
 template < typename targetidentifierT >
 void
-STDPConnectionHom< targetidentifierT >::get_status( DictionaryDatum& d ) const
+stdp_synapse< targetidentifierT >::get_status( DictionaryDatum& d ) const
 {
-
-  // base class properties, different for individual synapse
   ConnectionBase::get_status( d );
   def< double >( d, names::weight, weight_ );
-
-  // own properties, different for individual synapse
-  def< double >( d, names::Kplus, Kplus_ );
+  def< double >( d, names::tau_plus, tau_plus_ );
+  def< double >( d, names::lambda, lambda_ );
+  def< double >( d, names::alpha, alpha_ );
+  def< double >( d, names::mu_plus, mu_plus_ );
+  def< double >( d, names::mu_minus, mu_minus_ );
+  def< double >( d, names::Wmax, Wmax_ );
   def< long >( d, names::size_of, sizeof( *this ) );
 }
 
 template < typename targetidentifierT >
 void
-STDPConnectionHom< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
+stdp_synapse< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
 {
-  // base class properties
   ConnectionBase::set_status( d, cm );
   updateValue< double >( d, names::weight, weight_ );
+  updateValue< double >( d, names::tau_plus, tau_plus_ );
+  updateValue< double >( d, names::lambda, lambda_ );
+  updateValue< double >( d, names::alpha, alpha_ );
+  updateValue< double >( d, names::mu_plus, mu_plus_ );
+  updateValue< double >( d, names::mu_minus, mu_minus_ );
+  updateValue< double >( d, names::Wmax, Wmax_ );
 
-  updateValue< double >( d, names::Kplus, Kplus_ );
+  // check if weight_ and Wmax_ has the same sign
+  if ( not( ( ( weight_ >= 0 ) - ( weight_ < 0 ) ) == ( ( Wmax_ >= 0 ) - ( Wmax_ < 0 ) ) ) )
+  {
+    throw BadProperty( "Weight and Wmax must have same sign." );
+  }
 }
 
 } // of namespace nest
 
-#endif // of #ifndef STDP_CONNECTION_HOM_H
+#endif // of #ifndef STDP_SYNAPSE_H
