+++ models/tsodyks_synapse_hom.h
@@ -1,5 +1,5 @@
 /*
- *  tsodyks_connection.h
+ *  tsodyks_synapse_hom.h
  *
  *  This file is part of NEST.
  *
@@ -20,14 +20,12 @@
  *
  */
 
-#ifndef TSODYKS_CONNECTION_H
-#define TSODYKS_CONNECTION_H
+#ifndef TSODYKS_SYNAPSE_HOM_H
+#define TSODYKS_SYNAPSE_HOM_H
 
-
-// C++ includes:
-#include <cmath>
 
 // Includes from nestkernel:
+#include "common_properties_hom_w.h"
 #include "connection.h"
 
 namespace nest
@@ -38,7 +36,7 @@ namespace nest
 Short description
 +++++++++++++++++
 
-Synapse type with short term plasticity
+Synapse type with short term plasticity using homogeneous parameters
 
 Description
 +++++++++++
@@ -89,8 +87,6 @@ an arbitrary postsynaptic effect depending on y(t).
 Parameters
 ++++++++++
 
-The following parameters can be set in the status dictionary:
-
 ========  ======  ========================================================
  U        real    Parameter determining the increase in u with each spike
                   [0,1]
@@ -102,6 +98,12 @@ The following parameters can be set in the status dictionary:
  y        real    Initial fraction of synaptic vesicles in the synaptic
                   cleft [0,1]
 ========  ======  ========================================================
+
+Remarks:
+
+The weight and the parameters U, tau_psc, tau_fac, and tau_rec are common to
+all synapses of the model and must be set using SetDefaults on the synapse
+model.
 
 References
 ++++++++++
@@ -118,33 +120,65 @@ SpikeEvent
 See also
 ++++++++
 
-stdp_synapse, static_synapse, iaf_psc_exp, iaf_tum_2000
+tsodyks_synapse, stdp_synapse_hom, static_synapse_hom_w
 
 EndUserDocs */
 
+/**
+ * Class containing the common properties for all synapses of type
+ * tsodyks_synapse_hom.
+ */
+
+class TsodyksHomCommonProperties : public CommonPropertiesHomW
+{
+
+public:
+  /**
+   * Default constructor.
+   * Sets all property values to defaults.
+   */
+  TsodyksHomCommonProperties();
+
+  /**
+   * Get all properties and put them into a dictionary.
+   */
+  void get_status( DictionaryDatum& d ) const;
+
+  /**
+   * Set properties from the values given in dictionary.
+   */
+  void set_status( const DictionaryDatum& d, ConnectorModel& cm );
+
+  double tau_psc_; //!< [ms] time constant of postsyn current
+  double tau_fac_; //!< [ms] time constant for fascilitation
+  double tau_rec_; //!< [ms] time constant for recovery
+  double U_;       //!< asymptotic value of probability of release
+};
+
+
 template < typename targetidentifierT >
-class TsodyksConnection : public Connection< targetidentifierT >
+class tsodyks_synapse_hom : public Connection< targetidentifierT >
 {
 public:
-  typedef CommonSynapseProperties CommonPropertiesType;
+  typedef TsodyksHomCommonProperties CommonPropertiesType;
   typedef Connection< targetidentifierT > ConnectionBase;
 
   /**
    * Default Constructor.
    * Sets default values for all parameters. Needed by GenericConnectorModel.
    */
-  TsodyksConnection();
+  tsodyks_synapse_hom();
 
   /**
-     * Copy constructor from a property object.
-     * Needs to be defined properly in order for GenericConnector to work.
-     */
-  TsodyksConnection( const TsodyksConnection& ) = default;
+   * Copy constructor from a property object.
+   * Needs to be defined properly in order for GenericConnector to work.
+   */
+  tsodyks_synapse_hom( const tsodyks_synapse_hom& ) = default;
 
   /**
    * Default Destructor.
    */
-  ~TsodyksConnection()
+  ~tsodyks_synapse_hom()
   {
   }
 
@@ -153,7 +187,6 @@ public:
   // functions are used. Since ConnectionBase depends on the template parameter,
   // they are not automatically found in the base class.
   using ConnectionBase::get_delay_steps;
-  using ConnectionBase::get_delay;
   using ConnectionBase::get_rport;
   using ConnectionBase::get_target;
 
@@ -172,7 +205,7 @@ public:
    * \param e The event to send
    * \param cp Common properties to all synapses (empty).
    */
-  void send( Event& e, thread t, const CommonSynapseProperties& cp );
+  void send( Event& e, thread t, const TsodyksHomCommonProperties& cp );
 
   class ConnTestDummyNode : public ConnTestDummyNodeBase
   {
@@ -195,17 +228,15 @@ public:
   }
 
   void
-  set_weight( double w )
+  set_weight( double )
   {
-    weight_ = w;
+    throw BadProperty(
+      "Setting of individual weights is not possible! The common weights can "
+      "be changed via "
+      "CopyModel()." );
   }
 
 private:
-  double weight_;
-  double tau_psc_;     //!< [ms] time constant of postsyn current
-  double tau_fac_;     //!< [ms] time constant for fascilitation
-  double tau_rec_;     //!< [ms] time constant for recovery
-  double U_;           //!< asymptotic value of probability of release
   double x_;           //!< amount of resources in recovered state
   double y_;           //!< amount of resources in active state
   double u_;           //!< actual probability of release
@@ -220,13 +251,10 @@ private:
  */
 template < typename targetidentifierT >
 inline void
-TsodyksConnection< targetidentifierT >::send( Event& e, thread t, const CommonSynapseProperties& )
+tsodyks_synapse_hom< targetidentifierT >::send( Event& e, thread t, const TsodyksHomCommonProperties& cp )
 {
   const double t_spike = e.get_stamp().get_ms();
   const double h = t_spike - t_lastspike_;
-
-  Node* target = get_target( t );
-
 
   // t_lastspike_ = 0 initially
   // this has no influence on the dynamics, IF y = z = 0 initially
@@ -234,11 +262,11 @@ TsodyksConnection< targetidentifierT >::send( Event& e, thread t, const CommonSy
 
   // propagator
   // TODO: use expm1 here instead, where applicable
-  double Puu = ( tau_fac_ == 0.0 ) ? 0.0 : std::exp( -h / tau_fac_ );
-  double Pyy = std::exp( -h / tau_psc_ );
-  double Pzz = std::exp( -h / tau_rec_ );
+  double Puu = ( cp.tau_fac_ == 0.0 ) ? 0.0 : std::exp( -h / cp.tau_fac_ );
+  double Pyy = std::exp( -h / cp.tau_psc_ );
+  double Pzz = std::exp( -h / cp.tau_rec_ );
 
-  double Pxy = ( ( Pzz - 1.0 ) * tau_rec_ - ( Pyy - 1.0 ) * tau_psc_ ) / ( tau_psc_ - tau_rec_ );
+  double Pxy = ( ( Pzz - 1.0 ) * cp.tau_rec_ - ( Pyy - 1.0 ) * cp.tau_psc_ ) / ( cp.tau_psc_ - cp.tau_rec_ );
   double Pxz = 1.0 - Pzz;
 
   double z = 1.0 - x_ - y_;
@@ -251,7 +279,7 @@ TsodyksConnection< targetidentifierT >::send( Event& e, thread t, const CommonSy
   y_ *= Pyy;
 
   // delta function u
-  u_ += U_ * ( 1.0 - u_ );
+  u_ += cp.U_ * ( 1.0 - u_ );
 
   // postsynaptic current step caused by incoming spike
   double delta_y_tsp = u_ * x_;
@@ -261,8 +289,8 @@ TsodyksConnection< targetidentifierT >::send( Event& e, thread t, const CommonSy
   y_ += delta_y_tsp;
 
 
-  e.set_receiver( *target );
-  e.set_weight( delta_y_tsp * weight_ );
+  e.set_receiver( *get_target( t ) );
+  e.set_weight( delta_y_tsp * cp.get_weight() );
   e.set_delay_steps( get_delay_steps() );
   e.set_rport( get_rport() );
   e();
@@ -271,13 +299,8 @@ TsodyksConnection< targetidentifierT >::send( Event& e, thread t, const CommonSy
 }
 
 template < typename targetidentifierT >
-TsodyksConnection< targetidentifierT >::TsodyksConnection()
+tsodyks_synapse_hom< targetidentifierT >::tsodyks_synapse_hom()
   : ConnectionBase()
-  , weight_( 1.0 )
-  , tau_psc_( 3.0 )
-  , tau_fac_( 0.0 )
-  , tau_rec_( 800.0 )
-  , U_( 0.5 )
   , x_( 1.0 )
   , y_( 0.0 )
   , u_( 0.0 )
@@ -287,28 +310,21 @@ TsodyksConnection< targetidentifierT >::TsodyksConnection()
 
 template < typename targetidentifierT >
 void
-TsodyksConnection< targetidentifierT >::get_status( DictionaryDatum& d ) const
+tsodyks_synapse_hom< targetidentifierT >::get_status( DictionaryDatum& d ) const
 {
   ConnectionBase::get_status( d );
-  def< double >( d, names::weight, weight_ );
 
-  def< double >( d, names::U, U_ );
-  def< double >( d, names::tau_psc, tau_psc_ );
-  def< double >( d, names::tau_rec, tau_rec_ );
-  def< double >( d, names::tau_fac, tau_fac_ );
   def< double >( d, names::x, x_ );
   def< double >( d, names::y, y_ );
   def< double >( d, names::u, u_ );
-  def< long >( d, names::size_of, sizeof( *this ) );
 }
 
 template < typename targetidentifierT >
 void
-TsodyksConnection< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
+tsodyks_synapse_hom< targetidentifierT >::set_status( const DictionaryDatum& d, ConnectorModel& cm )
 {
   // Handle parameters that may throw an exception first, so we can leave the
-  // synapse untouched
-  // in case of invalid parameter values
+  // synapse untouched in case of invalid parameter values
   double x = x_;
   double y = y_;
   updateValue< double >( d, names::x, x );
@@ -323,35 +339,10 @@ TsodyksConnection< targetidentifierT >::set_status( const DictionaryDatum& d, Co
   y_ = y;
 
   ConnectionBase::set_status( d, cm );
-  updateValue< double >( d, names::weight, weight_ );
-
-  updateValue< double >( d, names::U, U_ );
-  if ( U_ > 1.0 || U_ < 0.0 )
-  {
-    throw BadProperty( "U must be in [0,1]." );
-  }
-
-  updateValue< double >( d, names::tau_psc, tau_psc_ );
-  if ( tau_psc_ <= 0.0 )
-  {
-    throw BadProperty( "tau_psc must be > 0." );
-  }
-
-  updateValue< double >( d, names::tau_rec, tau_rec_ );
-  if ( tau_rec_ <= 0.0 )
-  {
-    throw BadProperty( "tau_rec must be > 0." );
-  }
-
-  updateValue< double >( d, names::tau_fac, tau_fac_ );
-  if ( tau_fac_ < 0.0 )
-  {
-    throw BadProperty( "tau_fac must be >= 0." );
-  }
 
   updateValue< double >( d, names::u, u_ );
 }
 
 } // namespace
 
-#endif // TSODYKS_CONNECTION_H
+#endif // TSODYKS_SYNAPSE_HOM_H
